+++ lib/elixir/lib/macro/env.ex
@@ -20,6 +20,7 @@ defmodule Macro.Env do
   * `functions` - a list of functions imported from each module
   * `macros` - a list of macros imported from each module
   * `context_modules` - a list of modules defined in the current context
+  * `macro_aliases` - a list of aliases defined inside the current macro
   """
 
   @type name_arity :: { atom, non_neg_integer }
@@ -32,12 +33,13 @@ defmodule Macro.Env do
   @type macros :: [{ module, [name_arity] }]
   @type context_modules :: [module]
 
-  fields = [:module, :file, :line, :function, :aliases,
-            :context, :requires, :functions, :macros, :context_modules]
+  fields = [:module, :file, :line, :function, :aliases, :context, :requires,
+            :functions, :macros, :context_modules, :macro_aliases]
 
   types  = quote do: [module: module, file: file, line: line,
     function: name_arity, aliases: aliases, requires: requires,
-    functions: functions, macros: macros, context_modules: context_modules]
+    functions: functions, macros: macros, context_modules: context_modules,
+    macro_aliases: aliases]
 
   Record.deffunctions(fields, __MODULE__)
   Record.deftypes(fields, types, __MODULE__)