+++ lib/elixir/test/elixir/record_test.exs
@@ -52,11 +52,26 @@ end
 defmodule RecordTest.Macros do
   defrecordp :_user, name: "José", age: 25
 
+  defmacro gen do
+    quote do
+      alias RecordTest.Macros.Nested
+
+      def this_works, do: RecordTest.Macros.Nested[]
+      def this_should_too, do: Nested[]
+    end
+  end
+
   defrecord Nested do
     def nested_record_alias?(Nested[]) do
       true
     end
+
+    defrecord NestedInNested, it_compiles: true
   end
+
+  # Ensure there is no conflict in a nested module
+  # named record.
+  defrecord Record, [a: 1, b: 2]
 
   def new() do
     _user()
@@ -93,6 +108,11 @@ end
 
 defmodule RecordTest do
   use ExUnit.Case, async: true
+
+  # Check the access from the generated macro works
+  # as expected. If it compiles, we are good to go.
+  require RecordTest.Macros
+  RecordTest.Macros.gen
 
   test :record_access_with_nil_keyword do
     record = RecordTest.DynamicName.new(a: nil)